Project 13: CarManager exceptions.

+++ Project13.Entity/Employee.cs
@@ -9,7 +9,7 @@ namespace Project13.Entity
     public class Employee : Person, IEntity
     {
         public int Salary { get; set; }
-        public int ContactPhone { get; set; }
+        public long ContactPhone { get; set; }
         public string Adress { get; set; }
         public int OffDayMonthly { get; set; }
     }
